+++ pcsx2/MTGS.cpp
@@ -20,6 +20,7 @@
 #include <wx/wx.h>
 
 #include "GS.h"
+#include "GS/GS.h"
 #include "Gif_Unit.h"
 #include "MTVU.h"
 #include "Elfheader.h"
@@ -319,7 +320,7 @@ void SysMtgsThread::ExecuteTaskInThread()
 					Gif_Path& path   = gifUnit.gifPath[tag.data[2]];
 					u32       offset = tag.data[0];
 					u32       size   = tag.data[1];
-					if (offset != ~0u) GSgifTransfer((u32*)&path.buffer[offset], size/16);
+					if (offset != ~0u) GSgifTransfer((u8*)&path.buffer[offset], size/16);
 					path.readAmount.fetch_sub(size, std::memory_order_acq_rel);
 					break;
 				}
@@ -339,7 +340,7 @@ void SysMtgsThread::ExecuteTaskInThread()
 #endif
 					Gif_Path& path   = gifUnit.gifPath[GIF_PATH_1];
 					GS_Packet gsPack = path.GetGSPacketMTVU(); // Get vu1 program's xgkick packet(s)
-					if (gsPack.size) GSgifTransfer((u32*)&path.buffer[gsPack.offset], gsPack.size/16);
+					if (gsPack.size) GSgifTransfer((u8*)&path.buffer[gsPack.offset], gsPack.size/16);
 					path.readAmount.fetch_sub(gsPack.size + gsPack.readAmount, std::memory_order_acq_rel);
 					path.PopGSPacketMTVU(); // Should be done last, for proper Gif_MTGS_Wait()
 					break;
@@ -431,12 +432,12 @@ void SysMtgsThread::ExecuteTaskInThread()
 
 						case GS_RINGTYPE_INIT_READ_FIFO1:
 							MTGS_LOG( "(MTGS Packet Read) ringtype=Fifo1" );
-							GSinitReadFIFO( (u64*)tag.pointer);
+							GSinitReadFIFO( (u8*)tag.pointer);
 						break;
 
 						case GS_RINGTYPE_INIT_READ_FIFO2:
 							MTGS_LOG( "(MTGS Packet Read) ringtype=Fifo2, size=%d", tag.data[0] );
-							GSinitReadFIFO2( (u64*)tag.pointer, tag.data[0]);
+							GSinitReadFIFO2( (u8*)tag.pointer, tag.data[0]);
 						break;
 
 #ifdef PCSX2_DEVBUILD
